Reorganize module layout some more
