+++ bytecode/src/repr/chunk.rs
@@ -1,4 +1,4 @@
-use crate::{instruction::Instruction, value::Value};
+use super::{instruction::Instruction, value::Value};
 
 #[derive(Debug)]
 pub struct Chunk {
